requirements.txt
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2
django-crispy-forms==1.10.0
django-formtools==2.2
django-import-export==2.5.0
django-pure-pagination==0.3.0
django-reversion==3.0.9
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.18.1
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
mysqlclient==2.0.3
odfpy==1.4.1
openpyxl==3.0.6
Pillow==8.1.0
PyMySQL==1.0.2
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.3
xlrd==2.0.1
XlsxWriter==1.3.7
xlwt==1.3.0
